cloud-run/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# GCP Services
google-cloud-firestore==2.13.1
google-cloud-storage==2.10.0
google-cloud-pubsub==2.18.4
google-cloud-logging==3.8.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Finance Data APIs (Finnhub primary, Alpha Vantage fallback, yfinance last resort)
finnhub-python>=2.4.19
alpha-vantage>=2.3.1
yfinance>=0.2.31

# Caching
cachetools>=5.0.0

# MCP SDK
mcp>=0.9.0

# Utilities
python-dateutil==2.8.2
httpx>=0.25.0
